+++ src/main/java/com/bitbank/entities/UsersEntity.java
@@ -13,7 +13,6 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.JoinTable;
 import jakarta.persistence.ManyToMany;
 import jakarta.persistence.Table;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -38,7 +37,10 @@ public class UsersEntity {
 	@Column(name = "password", columnDefinition = "VARCHAR(256)")
 	private String password;
 
-	@ManyToMany(fetch = FetchType.LAZY)
+	@Column(name = "secret_mfa", columnDefinition = "VARCHAR(256)")
+	private String secretMfa;
+
+	@ManyToMany(fetch = FetchType.EAGER)
 	@JoinTable(name = "users_roles", joinColumns = @JoinColumn(name = "users_id"), inverseJoinColumns = @JoinColumn(name = "roles_id"))
 	@Builder.Default
 	private Set<RolesEntity> rolesEntity = new HashSet<>();